rename: mutation 훅 mutation postfix 일괄 삭제 및 적용
