+++ src/pages/FeedDetail/hooks/mutations/useDeleteLikeFeed.ts
@@ -3,23 +3,23 @@ import { AxiosError } from 'axios';
 
 import { http } from '../../../../api/http';
 
-const _postLike = (id: string) => {
-  return http.post(`/ideas/likes/${id}`);
+const _deleteLike = (id: string) => {
+  return http.delete(`/ideas/likes/${id}`);
 };
 
-const usePostLikeMutation = (id: string) => {
+const useDeleteLikeFeed = (id: string) => {
   const queryClient = useQueryClient();
-  const { mutate: postLike, ...rest } = useMutation({
-    mutationFn: _postLike,
+  const { mutate: deleteLike, ...rest } = useMutation({
+    mutationFn: _deleteLike,
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['feed', 'detail', id] });
     },
     onError: (error: AxiosError<{ message: string }>) => {
-      alert(error.response?.data.message ?? '좋아요에 실패했습니다.');
+      alert(error.response?.data.message ?? '좋아요 해제에 실패했습니다.');
     },
   });
 
-  return { postLike, ...rest };
+  return { deleteLike, ...rest };
 };
 
-export default usePostLikeMutation;
+export default useDeleteLikeFeed;
