+++ src/pages/FeedDetail/hooks/mutations/useDeleteScrapFeed.ts
@@ -3,23 +3,23 @@ import { AxiosError } from 'axios';
 
 import { http } from '../../../../api/http';
 
-const _deleteLike = (id: string) => {
-  return http.delete(`/ideas/likes/${id}`);
+const _deleteScrap = (id: string) => {
+  return http.delete<void>(`/bookmark/${id}`);
 };
 
-const useDeleteLikeMutation = (id: string) => {
+const useDeleteScrapFeed = (id: string) => {
   const queryClient = useQueryClient();
-  const { mutate: deleteLike, ...rest } = useMutation({
-    mutationFn: _deleteLike,
+  const { mutate: deleteScrap, ...rest } = useMutation({
+    mutationFn: _deleteScrap,
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['feed', 'detail', id] });
     },
     onError: (error: AxiosError<{ message: string }>) => {
-      alert(error.response?.data.message ?? '좋아요 해제에 실패했습니다.');
+      alert(error.response?.data.message ?? '스크랩 해제에 실패했습니다.');
     },
   });
 
-  return { deleteLike, ...rest };
+  return { deleteScrap, ...rest };
 };
 
-export default useDeleteLikeMutation;
+export default useDeleteScrapFeed;
