+++ src/pages/FeedDetail/hooks/mutations/usePostLikeFeed.ts
@@ -3,23 +3,23 @@ import { AxiosError } from 'axios';
 
 import { http } from '../../../../api/http';
 
-const _postScrap = (id: string) => {
-  return http.post<void>(`/bookmark/${id}`);
+const _postLike = (id: string) => {
+  return http.post(`/ideas/likes/${id}`);
 };
 
-const usePostScrapMutation = (id: string) => {
+const usePostLike = (id: string) => {
   const queryClient = useQueryClient();
-  const { mutate: postScrap, ...rest } = useMutation({
-    mutationFn: _postScrap,
+  const { mutate: postLike, ...rest } = useMutation({
+    mutationFn: _postLike,
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['feed', 'detail', id] });
     },
     onError: (error: AxiosError<{ message: string }>) => {
-      alert(error.response?.data.message ?? '스크랩에 실패했습니다.');
+      alert(error.response?.data.message ?? '좋아요에 실패했습니다.');
     },
   });
 
-  return { postScrap, ...rest };
+  return { postLike, ...rest };
 };
 
-export default usePostScrapMutation;
+export default usePostLike;
